Set up background service for account retrieval networking
